Final Updates 05/31/23

+++ app/Http/Controllers/ListingController.php
@@ -49,6 +49,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logo', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing Created Successfully!');
@@ -63,6 +65,10 @@ class ListingController extends Controller
     //update listin data
     public function update(Request $request, Listing $listing)
     {
+        //make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         // dd($request->all());
         $formFields = $request->validate([
             'title' => 'required',
@@ -86,8 +92,18 @@ class ListingController extends Controller
     //Delete listing
     public function destroy(Listing $listing)
     {
+        //make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing Deleted Successfully!');
+    }
+
+    //Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
